+++ crates/log-utils/src/lib.rs
@@ -70,7 +70,8 @@ fn default_directives() -> Vec<Directive> {
     namespaces
         .into_iter()
         .map(|ns| {
-            ns.parse()
+            ns.trim()
+                .parse()
                 .unwrap_or_else(|e| panic!("cannot parse {ns} to Directive: {e}"))
         })
         .collect()
@@ -143,7 +144,8 @@ pub fn enable_logs_for(spec: LogSpec) {
 
     let mut filter = tracing_subscriber::EnvFilter::builder()
         .with_default_directive(spec.level.into())
-        .from_env_lossy();
+        .from_env()
+        .expect("invalid RUST_LOG");
 
     for d in spec.directives {
         filter = filter.add_directive(d);