+++ nox/src/behaviour/identify.rs
@@ -56,10 +56,11 @@ impl FluenceNetworkBehaviour {
                 }
 
                 if supports_fluence {
+                    let protocols: Vec<_> = info.protocols.iter().map(|p| p.to_string()).collect();
                     log::debug!(
                         target: "network",
                         "Found fluence peer {}: protocols: {:?} version: {} listen addrs {:?}",
-                        peer_id, info.protocols, info.protocol_version, addresses
+                        peer_id, protocols, info.protocol_version, addresses
                     );
                     // Add addresses to connection pool disregarding whether it supports kademlia or not
                     // we want to have full info on non-kademlia peers as well
@@ -73,8 +74,8 @@ impl FluenceNetworkBehaviour {
                         target: "blocked",
                         "Found peer {} not supported fluence protocol, protocols: {:?} version: {} listen addrs {:?}. skipping...",
                         peer_id, info.protocols,
-                    info.protocol_version,
-                    info.listen_addrs
+                        info.protocol_version,
+                        info.listen_addrs
                     );
                     let (out, _inlet) = oneshot::channel();
                     self.connection_pool.disconnect(peer_id, out);