+++ particle-builtins/src/builtins.rs
@@ -215,6 +215,7 @@ where
             ("srv", "create") => wrap(self.create_service(args, particle)),
             ("srv", "get_interface") => wrap(self.get_interface(args, particle)),
             ("srv", "resolve_alias") => wrap(self.resolve_alias(args, particle)),
+            ("srv", "resolve_alias_opt") => wrap(self.resolve_alias_opt(args, particle)),
             ("srv", "add_alias") => wrap_unit(self.add_alias(args, particle)),
             ("srv", "remove") => wrap_unit(self.remove_service(args, particle)),
             ("srv", "info") => wrap(self.get_service_info(args, particle)),
@@ -295,7 +296,18 @@ where
 
             ("vault", "put") => wrap(self.vault_put(args, particle)),
             ("vault", "cat") => wrap(self.vault_cat(args, particle)),
-            ("run-console", "print")          => wrap_unit(Ok(log::debug!(target: "run-console", "{}", json!(args.function_args)))),
+            ("run-console", "print") => {
+                let function_args = args.function_args.iter();
+                let decider = function_args.filter_map(JValue::as_str).any(|s| s.contains("decider"));
+                if decider {
+                    // if log comes from decider, log it as INFO
+                    log::info!(target: "run-console", "{}", json!(args.function_args));
+                } else {
+                    // log everything else as DEBUG
+                    log::debug!(target: "run-console", "{}", json!(args.function_args));
+                }
+                wrap_unit(Ok(()))
+            },
 
             _ => FunctionOutcome::NotDefined { args, params: particle },
         }
@@ -848,7 +860,7 @@ where
         self.services.remove_service(
             &params.id,
             params.host_id,
-            service_id_or_alias,
+            &service_id_or_alias,
             params.init_peer_id,
             false,
         )?;
@@ -889,6 +901,18 @@ where
             .resolve_alias(&params.id, params.host_id, alias)?;
 
         Ok(JValue::String(service_id))
+    }
+
+    fn resolve_alias_opt(&self, args: Args, params: ParticleParams) -> Result<JValue, JError> {
+        let mut args = args.function_args.into_iter();
+        let alias: String = Args::next("alias", &mut args)?;
+        let service_id_opt = self
+            .services
+            .resolve_alias(&params.id, params.host_id, alias)
+            .map(|id| vec![JValue::String(id)])
+            .unwrap_or(vec![]);
+
+        Ok(Array(service_id_opt))
     }
 
     fn get_service_info(&self, args: Args, params: ParticleParams) -> Result<JValue, JError> {