+++ sorcerer/src/spell_builtins.rs
@@ -13,11 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-use fluence_spell_dtos::value::{ScriptValue, SpellValueT, StringValue, UnitValue};
 use serde_json::{json, Value as JValue, Value, Value::Array};
 
-use crate::utils::{parse_spell_id_from, process_func_outcome};
-use fluence_spell_dtos::trigger_config::{TriggerConfig, TriggerConfigValue};
+use crate::utils::parse_spell_id_from;
+use fluence_spell_dtos::trigger_config::TriggerConfig;
 use key_manager::KeyManager;
 use libp2p::PeerId;
 use particle_args::{Args, JError};
@@ -25,6 +24,7 @@ use particle_execution::ParticleParams;
 use particle_services::{ParticleAppServices, ServiceType};
 use spell_event_bus::api::EventBusError;
 use spell_event_bus::{api, api::SpellEventBusApi};
+use spell_service_api::{CallParams, SpellServiceApi};
 use spell_storage::SpellStorage;
 use std::time::Duration;
 
@@ -33,10 +33,10 @@ pub async fn remove_spell(
     spell_storage: &SpellStorage,
     services: &ParticleAppServices,
     spell_event_bus_api: &SpellEventBusApi,
-    spell_id: String,
+    spell_id: &str,
     worker_id: PeerId,
 ) -> Result<(), JError> {
-    if let Err(err) = spell_event_bus_api.unsubscribe(spell_id.clone()).await {
+    if let Err(err) = spell_event_bus_api.unsubscribe(spell_id.to_string()).await {
         log::warn!(
             "can't unsubscribe a spell {spell_id} from its triggers via spell-event-bus-api: {err}"
         );
@@ -45,7 +45,7 @@ pub async fn remove_spell(
         )));
     }
 
-    spell_storage.unregister_spell(worker_id, &spell_id);
+    spell_storage.unregister_spell(worker_id, spell_id);
     services.remove_service(particle_id, worker_id, spell_id, worker_id, true)?;
     Ok(())
 }
@@ -55,6 +55,7 @@ pub async fn install_spell(
     services: &ParticleAppServices,
     spell_storage: &SpellStorage,
     spell_event_bus_api: &SpellEventBusApi,
+    spell_service_api: &SpellServiceApi,
     worker_id: PeerId,
     particle_id: String,
     ttl: u64,
@@ -72,51 +73,13 @@ pub async fn install_spell(
     )?;
     spell_storage.register_spell(worker_id, spell_id.clone());
 
-    // TODO: refactor these service calls
+    let params = CallParams::local(spell_id.clone(), worker_id, Duration::from_millis(ttl));
     // Save the script to the spell
-    process_func_outcome::<UnitValue>(
-        services.call_function(
-            worker_id,
-            &spell_id,
-            "set_script_source_to_file",
-            vec![json!(script)],
-            None,
-            worker_id,
-            Duration::from_millis(ttl),
-        ),
-        &spell_id,
-        "set_script_source_to_file",
-    )?;
-
+    spell_service_api.set_script(params.clone(), script)?;
     // Save init_data to the spell's KV
-    process_func_outcome::<UnitValue>(
-        services.call_function(
-            worker_id,
-            &spell_id,
-            "set_json_fields",
-            vec![json!(init_data.to_string())],
-            None,
-            worker_id,
-            Duration::from_millis(ttl),
-        ),
-        &spell_id,
-        "set_json_fields",
-    )?;
-
+    spell_service_api.update_kv(params.clone(), init_data)?;
     // Save trigger config
-    process_func_outcome::<UnitValue>(
-        services.call_function(
-            worker_id,
-            &spell_id,
-            "set_trigger_config",
-            vec![json!(user_config)],
-            None,
-            worker_id,
-            Duration::from_millis(ttl),
-        ),
-        &spell_id,
-        "set_trigger_config",
-    )?;
+    spell_service_api.set_trigger_config(params, user_config)?;
 
     if let Some(config) = config {
         // Scheduling the spell
@@ -127,7 +90,7 @@ pub async fn install_spell(
             log::warn!("can't subscribe a spell {} to triggers {:?} via spell-event-bus-api: {}. Removing created spell service...", spell_id, config, err);
 
             spell_storage.unregister_spell(worker_id, &spell_id);
-            services.remove_service(&particle_id, worker_id, spell_id, worker_id, true)?;
+            services.remove_service(&particle_id, worker_id, &spell_id, worker_id, true)?;
 
             return Err(JError::new(format!(
                 "can't install a spell due to an internal error while subscribing to the triggers: {err}"
@@ -151,53 +114,23 @@ pub struct SpellInfo {
 }
 
 pub fn get_spell_info(
-    services: &ParticleAppServices,
+    spell_service_api: &SpellServiceApi,
     worker_id: PeerId,
     ttl: u64,
     spell_id: String,
 ) -> Result<SpellInfo, JError> {
-    let trigger_config_value = process_func_outcome::<TriggerConfigValue>(
-        services.call_function(
-            worker_id,
-            &spell_id,
-            "get_trigger_config",
-            vec![],
-            None,
-            worker_id,
-            Duration::from_millis(ttl),
-        ),
-        &spell_id,
-        "get_trigger_config",
-    )
-    .map_err(|e| JError::new(f!("Failed to get trigger_config for spell {spell_id}: {e}")))?;
-    let trigger_config = if trigger_config_value.is_success() {
-        trigger_config_value.config
-    } else {
-        return Err(JError::new(trigger_config_value.error));
-    };
+    let params = CallParams::local(spell_id.clone(), worker_id, Duration::from_millis(ttl));
+    let trigger_config = spell_service_api
+        .get_trigger_config(params.clone())
+        .map_err(|e| JError::new(f!("Failed to get trigger_config for spell {spell_id}: {e}")))?;
 
-    let script_value = process_func_outcome::<ScriptValue>(
-        services.call_function(
-            worker_id,
-            &spell_id,
-            "get_script_source_from_file",
-            vec![],
-            None,
-            worker_id,
-            Duration::from_millis(ttl),
-        ),
-        &spell_id,
-        "get_script_source_from_file",
-    )
-    .map_err(|e| JError::new(f!("Failed to get trigger_config for spell {spell_id}: {e}")))?;
-    if script_value.is_success() {
-        Ok(SpellInfo {
-            script: script_value.source_code,
-            trigger_config,
-        })
-    } else {
-        Err(JError::new(script_value.error))
-    }
+    let script = spell_service_api
+        .get_script(params)
+        .map_err(|e| JError::new(f!("Failed to get trigger_config for spell {spell_id}: {e}")))?;
+    Ok(SpellInfo {
+        script,
+        trigger_config,
+    })
 }
 
 pub(crate) async fn spell_install(
@@ -206,12 +139,15 @@ pub(crate) async fn spell_install(
     spell_storage: SpellStorage,
     services: ParticleAppServices,
     spell_event_bus_api: SpellEventBusApi,
+    spell_service_api: SpellServiceApi,
     key_manager: KeyManager,
 ) -> Result<JValue, JError> {
     let mut args = sargs.function_args.clone().into_iter();
     let script: String = Args::next("script", &mut args)?;
     let init_data: JValue = Args::next("data", &mut args)?;
-    let user_config: TriggerConfig = Args::next("config", &mut args)?;
+    let trigger_config: TriggerConfig = Args::next("trigger_config", &mut args)?;
+    let alias: Option<String> = Args::next_opt("alias", &mut args)?;
+
     let init_peer_id = params.init_peer_id;
 
     let is_management = key_manager.is_management(init_peer_id);
@@ -232,14 +168,36 @@ pub(crate) async fn spell_install(
         &services,
         &spell_storage,
         &spell_event_bus_api,
+        &spell_service_api,
         worker_id,
-        params.id,
+        params.id.clone(),
         params.ttl as u64,
-        user_config,
+        trigger_config,
         script,
         init_data,
     )
     .await?;
+
+    if let Some(alias) = alias {
+        if let Err(e) = services.add_alias(alias.clone(), worker_id, spell_id.clone(), worker_id) {
+            // Remove the spell if we failed to add an alias
+            remove_spell(
+                &params.id,
+                &spell_storage,
+                &services,
+                &spell_event_bus_api,
+                &spell_id,
+                worker_id,
+            )
+            .await?;
+
+            return Err(JError::new(format!(
+                "Failed to add alias {} for spell {}: {:?}",
+                alias, spell_id, e
+            )));
+        }
+    }
+
     Ok(JValue::String(spell_id))
 }
 
@@ -291,7 +249,7 @@ pub(crate) async fn spell_remove(
         &spell_storage,
         &services,
         &spell_event_bus_api,
-        spell_id,
+        &spell_id,
         worker_id,
     )
     .await
@@ -302,6 +260,7 @@ pub(crate) async fn spell_update_config(
     params: ParticleParams,
     services: ParticleAppServices,
     spell_event_bus_api: SpellEventBusApi,
+    spell_service_api: SpellServiceApi,
     key_manager: KeyManager,
 ) -> Result<(), JError> {
     let mut args = args.function_args.into_iter();
@@ -326,19 +285,12 @@ pub(crate) async fn spell_update_config(
     let user_config: TriggerConfig = Args::next("config", &mut args)?;
     let config = api::from_user_config(&user_config)?;
 
-    process_func_outcome::<UnitValue>(
-        services.call_function(
-            worker_id,
-            &spell_id,
-            "set_trigger_config",
-            vec![json!(user_config)],
-            None,
-            worker_id,
-            Duration::from_millis(params.ttl as u64),
-        ),
-        &spell_id,
-        "set_trigger_config",
-    )?;
+    let params = CallParams::local(
+        spell_id.clone(),
+        worker_id,
+        Duration::from_millis(params.ttl as u64),
+    );
+    spell_service_api.set_trigger_config(params, user_config)?;
 
     let result: Result<(), EventBusError> = try {
         // we unsubscribe the spell from the current config anyway
@@ -370,29 +322,20 @@ pub(crate) fn get_spell_id(params: ParticleParams) -> Result<JValue, JError> {
 pub(crate) fn get_spell_arg(
     args: Args,
     params: ParticleParams,
-    services: ParticleAppServices,
+    spell_service_api: SpellServiceApi,
 ) -> Result<JValue, JError> {
     let spell_id = parse_spell_id_from(&params)?;
     let key = args.function_name;
+    let call_params = CallParams::from(spell_id.clone(), params);
 
-    let str_value = process_func_outcome::<StringValue>(
-        services.call_function(
-            params.host_id,
-            &spell_id,
-            "get_string",
-            vec![json!(key)],
-            Some(params.id.clone()),
-            params.init_peer_id,
-            Duration::from_millis(params.ttl as u64),
-        ),
-        &spell_id,
-        "get_string",
-    )
-    .map_err(|e| JError::new(f!("Failed to get argument {key} for spell {spell_id}: {e}")))?;
+    let str_value = spell_service_api
+        .get_string(call_params, key.clone())
+        .map_err(|e| JError::new(f!("Failed to get argument {key} for spell {spell_id}: {e}")))
+        .and_then(|value| value.ok_or_else(|| JError::new("value not found")))?;
 
-    serde_json::from_str(&str_value.str).map_err(|e| {
+    serde_json::from_str(&str_value).map_err(|e| {
         JError::new(f!(
-            "Failed to parse argument {key} {str_value.str} ${str_value.error} for spell {spell_id}: {e}"
+            "Failed to parse argument `{key} -> {str_value}` for spell {spell_id}: {e}"
         ))
     })
 }
@@ -400,57 +343,35 @@ pub(crate) fn get_spell_arg(
 pub(crate) fn store_error(
     mut args: Args,
     params: ParticleParams,
-    services: ParticleAppServices,
+    spell_service_api: SpellServiceApi,
 ) -> Result<(), JError> {
     let spell_id = parse_spell_id_from(&params)?;
 
     args.function_args.push(json!(params.timestamp));
-    process_func_outcome::<UnitValue>(
-        services.call_function(
-            params.host_id,
-            &spell_id,
-            "store_error",
-            args.function_args.clone(),
-            Some(params.id.clone()),
-            params.init_peer_id,
-            Duration::from_millis(params.ttl as u64),
-        ),
-        &spell_id,
-        "store_error",
-    )
-    .map(drop)
-    .map_err(|e| {
-        JError::new(format!(
-            "Failed to store error {:?} for spell {}: {}",
-            args.function_args, spell_id, e
-        ))
-    })
+    let call_params = CallParams::from(spell_id.clone(), params);
+    spell_service_api
+        .store_error(call_params, args.function_args.clone())
+        .map_err(|e| {
+            JError::new(format!(
+                "Failed to store error {:?} for spell {}: {}",
+                args.function_args, spell_id, e
+            ))
+        })
 }
 
 pub(crate) fn store_response(
     args: Args,
     params: ParticleParams,
-    services: ParticleAppServices,
+    spell_service_api: SpellServiceApi,
 ) -> Result<(), JError> {
     let spell_id = parse_spell_id_from(&params)?;
     let response: JValue = Args::next("spell_id", &mut args.function_args.into_iter())?;
-    process_func_outcome::<UnitValue>(
-        services.call_function(
-            params.host_id,
-            &spell_id,
-            "set_json_fields",
-            vec![json!(response.to_string())],
-            Some(params.id.clone()),
-            params.init_peer_id,
-            Duration::from_millis(params.ttl as u64),
-        ),
-        &spell_id,
-        "set_json_fields",
-    )
-    .map(drop)
-    .map_err(|e| {
-        JError::new(format!(
-            "Failed to store response {response} for spell {spell_id}: {e}"
-        ))
-    })
+    let call_params = CallParams::from(spell_id.clone(), params);
+    spell_service_api
+        .update_kv(call_params, response.clone())
+        .map_err(|err| {
+            JError::new(format!(
+                "Failed to store response {response} for spell {spell_id}: {err}"
+            ))
+        })
 }